compost/osint/requirements.txt
telethon>=1.36
openai>=1.0
python-dotenv>=1.0
pydantic>=2.0
